Speed up TCP Server initialisation
Previously the TCP Server would only start after parts of Puppet were
initialised.  This commit changes the startup behaviour so that the TCP Server
is running as soon as the Puppet Version can be determined, and further
initialisation is handed to a different thread.  A new command line option of
--start-slow is added, mainly for testing, to slow down the startup if required.

This will help when the server is started by the VSCode client directly so that
the TCP socket is available as soon as possible.