(GH-54) Fix Puppet Resource Command
Previously the output from the Puppet Resource command in the language server
did not match the output from a normal `puppet resource` invocation.  This
commit modifies the language server to mimic the behaviour from puppet-agent
by use the prune_parameters method prior to send the data back to the client.
Also and additional newline is placed between each resource for better
readability.