(GH-7) Start Language server without a wait
This commit starts the language server without a wait/sleep using stdio
and events on the process created for the language server. It also
introduces a 'Puppet' outputchannel for logging that the user can
see.

This regresses windows support, as long running process creation on
windows is failing. This will be addressed in another commit.