(GH-57) Add ability to use specific version of Puppet Gem
Previously it was not possible to test on older puppet gems as the gemfile
always defaulted to latest.  This commit adds the PUPPET_GEM_VERSION
environment variable which can be used to set the version of the puppet gem
during a bundle.  This is the same workflow as regular Puppet modules.