Add a version to the language server
Previosuly the Language Server had no concept of a version.  While for local
operations this is ok, but when using the server in a non-local way, it may
mean that the client is talking to an "older" or "younger" server version.  This
commit adds the concept of a language server version to the server and message
protocol so that the language server can communicate its version.  It is
expected that this will be leveraged in later commits.