(GH-65) Add Windows support to local Language Server
This commit adds Windows support to the local Language Server. It detects when
a 32bit VS Code is on a 64bit OS and will use the appropriate file system
redirection to find a Puppet Agent installation.  This commit also improves the
user experience and will not throw large errors, but instead warns the user that
the language server could not be started and has reduced functionality.