Update extension to honor local language server settings
This commit updates the local language server invocation to honor the extension
settings, for example, port or debug file.  This commit removes the
stopOnClientExit property as it does not make sense on the local language server
and will cause issues on extension restart.  This commit adds a debugFilePath
setting for the local language server to direct all debug output to this file.