(GH-98) Add test for function loading and puppet fixtures
This commit adds tests for loading functions with errors.  This commit adds the
required file structure and fixture setting for a custom puppet environment.
This is needed to isolate the spec test fixtures from a potentially
misconfigured local puppet agent environment.  Resulting in possible false
negative and positive test results.