(GH-98) Fix Type and Function loading
Previously type and function loading would prematurely fail and stop the
language server if a type or function load threw an error.  This appears to be
a facet of the Puppet loadall function.  This commit changes the loading to
mimic the loadall function but instead, ignore any loading errors and continue.
This commit also adds some debug messages after types and function have
completed loading which is useful during debugging.

This commit also fixes function loader which was not idepmotent and was causing
functions to be loaded mulitple times.

This commit also fixes a typo in the reset function.