(GH-98) Fix puppet initialization in Language Server
Previously the init_puppet_worker was loading the function environment twice
which caused the Language Server to start slowly.  This commit removes the
unneeded loading.

This commit also adds the ability to send Puppet Settings on initialization.
This will only be used for test purposes and is not surfaced throught the CLI
options.