(GH-125) Honor inline puppet lint directives
Previously inline linting directives were not being honored by the document
validator.  This was found due to not calling the load_data method prior to
running the lint process.  This commit calls the load_data method prior to
linting and then adds a check to ignore any linting errors that were ignored by
puppet-lint.  This commit also adds tests for this scenario.