(GH-120) Fix default PuppetAgentDir on Windows
Previously the PuppetAgentDir default value was set to 'normal' which appears
to be a copy-paste error.  This causes the extension to try and load the
puppet language server at 'normal' in the file system which then promptly fails
on Windows.  This commit changes the default to null in JSON which then
appears in the TypeScript as undefined, and restores the previous behaviour.