(GH-130) Prepare to use PDK ruby if puppet agent not available
Previously the language server required a puppet-agent installation however
some users are not able to install a puppet-agent, but instead the PDK.  This
commit changes the connect.ts script;

- Refactors the ruby detection into two helpers, getLanguageServerFromPDK and
  getLanguageServerFromPuppetAgent.  These helpers are responsible for detecting
  and setting up the process creation for their respective ruby environments on
  any platform type
- Removes the need for environment.bat on Windows by setting up the process env
  variables at process creation
- Updates the logging to be more verbose in the detection methods

Note the functionality to actually use the PDK is currently disabled as the
configuration settings and user story need to be completed first.  However to
not lose the effort, the code still exists but is commented out.