Fix incorrect logger when a client error occurs
Previously, when the client failed to connect to the Language Server, e.g. the
local language server crashes, the UI was not updated to say that it failed.
This was due to an error being thrown in the error handler.  This commit changes
the connection handler to use the reference to the language client instead of,
`this` which is a Socket object.