(GH-61) Add click menu to Puppet status bar
Previously there was no way to restart the Language Client or view the Puppet
Client logs easily.  This commit adds a click menu which displays an option to
either restart the Lanugage Client or view the connection logs.  The location
to click is on the Language Server status bar item in the bottom right corner.