(GH-144) Show All PDK Commands
This commit unhides all pdk commands regardless if vscode thinks a
puppet file is open or not.

When in a window that hasn't opened a Puppet file (pp or epp), the PDK
commands are not available. This was an attempt to not show commands
that can't run without being inside a PDK module directory, but has a
side effect that after running 'pdk new module' you can't use the rest
of the pdk commands without having had run one of the commands.

This removes hiding the commands until we can research how to use the
when clause better.