(maint) Fix Connection Restart
Previously the connection stop method was stopping the language client and then
killing the Language Server process, however this caused a timing issue as the
client was trying to continually send a Shutdown event.  This then caused a log
storm of failed TCP events trying to send the Shutdown event to the non-existant
Server process.  This commit removes the Language Server process killing as the
spawned Server Process, by default, will automatically close when the client
disconnects.