(GH-169) Fix colourisation of comments in class params
Previously the sytnax highlighter would not highlight comments in the class
parameters section correctly.  This was compounded further if an apostrophe
was used in the comments, the highlighter would then mark the remainder of the
document as a string.  This commit adds the line_comment pattern to be available
when inside a class (or node) header definition.  The header is everything from
the word `class` to the first brace `{`.