(GH-177) Remove version file detection
This commit removes the check for the VERSION file inside the puppet
agent install path. This was originally used as a check that you had a
valid Puppet Agent path specified, but since there are variances whether
the file exists or not or where it's located in each package for each
platform, and it doesn't provide any actual value, it makes sense to
remove it.