(maint) Do not error in validation exception handler
Previosuly if an unexpected type of error came through the document validator
it would itself raise a new exception and crash the Language Server.  This was
caused by the lack of a :line or :pos method on the exception or the inner
exception.  This commit changes the detection logic to ignore any exception that
does not contain line and position information in the exception object.  These
can be ignored as they are not Puppet manifest errors but a symptom of something
else.