(maint) Use noop by default and fix readme typo
Previously the debug server would default to applying changes to the system
instead of being opt-in.  This commit changes the deafault value of noop to true
to make it safe by default.  This commit also fixes typos in the README and
credits Corey Osman with ideas behind the debug server.