(GH-204) Split tests due to monkey patching conflicts
Previously it was hoped to have the tests for the debug and language servers in
the same directory however as the two applications are separate and have very
distinct setups, they needed to be split.  This commit splits out the language
server tests into their own directory and modifies the Rakefile to be able to
invoke either set of tests.

Note the --default-path parameter will now be needed when running rspec directly
for example;

rspec --default-path ./spec/debugserver spec/debugserver/unit

This commit also modifies the appveyor.yaml file to run both the language server
and debug server tests (yet to be written) in all modern puppet gem versions.