(maint) Add integration tests for Debug Server
Previously the test coverage on the debug server was minimal.  This commit adds
integration tests for the debug server.  In particular it starts a debug server
as a separate process and emulates debug client TCP messages.  The kitchen sink
test file, tests all of the features of the debug server.