(maint) Suspend evaluation hooks when paused
Previously the evaluate hooks would be firing when the debug session was paused.
This means when doing string evaluations the POPS evaluation depth would be
modified and potentially other hooks (function, exception) would be firing in
error.  This commit guards the evaluation hooks to exit early if the session is
paused.

This commit also fixes a bug in the after evaluation hook which was decrementing
the counter back to zero, which then broke the Step Out functionality.  This
commit changes the functionality to only decrement by one per invocation.