(maint) Add support for stack frames in Puppet 4
Previously, generating the stack frame list was raising errors due to Puppet 4
not exposing the same methods.  This was mostly fixed in 6740faf2eb185 however
the stack frames were not testable.  This commit changes the location detection
to use the same helper methods created in 6740faf2eb185 and also adds the
pos_on_line helper method.