(maint) Update extension to latest template
This commit uses the extension generated by `yo code` as an exmaple to
update this extension to the latest structure recommended by VS Code.

The tasks, settings, and launch json files that control extension
development and testing have been updated to follow the latest versions.

The typescript and tslint settings have been updated to the latest
versions, and their config files updated to the latest standards. Strict
syntax checking was enabled, future work will bring the files into
compliance, then syntax errors will be turned into errors in the build
process.

The build process (pacakge.json and gulpfile.js) have been modified to
move inline with latest practices, however this breaks building the
extension. This is broken already by the multi-part move we are doing
to seperate the client extension from the language server, but this
commit breaks it further by removing the gulp commands. This is to lay
the ground work to making the build use gulp-typescript commands.