(GH-2410) Honor specified tcp port
This commit fixes the logic that determines the port used to connect
to the LanguageServer if using the `tcp` protocol. Previously it would
always pick a random unused port to start and connect to the language
server, now it will use the port specified in the configuration if it is
there.

Note, the default value for the `puppet.languageserver.port`
configuration value was removed in this commit. If it was left, there
would always be a value for the port, and the logic would never choose
a random port. This way, the default is actually random as the original
behavior implies but can be set to determinisitic if the user requires.