(GH-2410) Honor specified tcp port
This commit fixes the logic that determines the port used to connect
to the LanguageServer if using the `tcp` protocol. Previously it would
always pick a random unused port to start and connect to the language
server, now it will use the port specified in the configuration if it is
there.