(GH-274) Remove random port detection
This commit removes the code that detects and uses a random port for the
LanguageServer to run on. It instead parses the STDOUT message from the
LanguageServer for the port it is running on and uses that to connect.
If the user sets a port in the config, that is used instead of reading
the port from the LanguageServer.