(GH-301) Fail fast if no agent installed
This commit adds logic to check if there is a valid Puppet Agent binary
file path configured at extension load. It will then throw an
informative error if there isn't a binary found so the user knows what
won't work without an agent.

Previously if the extension couldn't find a Puppet Agent installation or
was configured with an incorrect path, the extension silently failed to
load the language server and there wasn't much indication you're in a
broken state.