(GH-317) Fix src folder path references
This commit corrects the relative path references to code not in the
`src` root directory so that when it is compiled into javascript, it is
in the correct path when either the tests are run or the extension is
loaded. In some cases this is more than a path reference change, and
actually changed how the type is referenced.