(GH-315) Update path resolution depending on installType
This commit updates puppetBaseDir to point to pdk or puppet depending
on installType. It relies on the new PathResolver class to make sure
we resolve the correct install root, as well as detection logic to
determine which platform we are on.

It also adds the PDK directories we will need to implement to the
interface to support PDK as a source.