(GH-353) Set Puppet Agent to default source
This commit sets the Puppet Agent as the default source for the
environment needed for the Language Server to run. This allows us to add
the PDK as a source and test it before making it the default for all
users

It also adjusts the tests to account for the new default of
puppet-agent.