(GH-355) Re-add Telemetry
This commit re-adds the version details to the telemetry implementation,
but not in the same exact place. This adds the capture to the closing
statement for the connection manager, which is called when the extension
stops. This allows us to send one event per 'session', and since we
include time spent we will be able to differentiate in appinsights.