(GH-278) Use a settings interface to abstract workspace settings
This commit creates the Settings interface and static methods which can parse
a workspace settings environment.  This also includes helpers for detecting
legacy or deprecated settings and can translate to the newer format.

This commit then implements the settings interface throughout the extension.

This commit also changes the ProtocolType enum to be a string instead of a
number.  This was needed due to the typecasting not translating strings into the
enum number.  By using a string, we can avoid this issue.