(maint) Rename the output channel for the Puppet Language Server
Previously the Puppet Langauge Server output channel would have a vague name of
`tcp lang server` with the host and ip address.  However now that we can use
different protocols, e.g. STDIO, this don't make as much sense.  This commit
changes the name which is more descriptive of what it's for.