(GH-400) Fix Path environment parsing for Ruby Helper
Previously the RubyHelper manipulates the path environment however due to Object
property names being case sensitive it could simply be that it's called Path or
path, particularly on Windows not so much on Linux etc.. It's also possible that
there is no PATH set but unlikely.

This commit changes the Path processing be to case-insensitive and force it to
be called PATH

This commit also removes environment variables whose value is set to null or
undefined.