(GH-323) Update the readme with gifs
This commit updates the readme with gifs that showcase individual
features. It removes the large all-in-one gif from the readme, as it's
replaced by the smaller and more targeted gifs. It also moves the
debugger gif to after the description so its before configuration
detail, which was also bulleted for easier reading.