(GH-422) Move from previewHTML to WebView API
Previously the Node Graph Feature use the previewHTML API within VSCode, however
this being actively deprecated by the VSCode team.  This commit wires up the
static Node Graph Preview to call the Language Server and request the DOT text
of the compiled manifest.  A difference in the new API is the HTML must be a
full HTML document.  The older API expected a partial document.