(GH-422) Colorize the Node Graph Preview correctly
Previously the Node Graph Feature used hardcoded colors for the graph, expecting
a dark theme.  This commit adds in CSS coloring to override the generated SVG
content and instead ise the VSCode theme coloring.  This means the preview will
now reflect the correct colors of the users theme, and, also auto-updates as the
user browses new themes.