(GH-412) Add ConnectionHandler
This commit adds a new ConnectionHandler abstract class that implements
all logic needed to create, configure, and start a `LanguageClient` to
connect with a `LanguageServer`. It does not handle setting up the logic
needed to create, configure, or start the protocol specific code needed
to communicate. That logic will be handled by classes that extend
`ConnectionHandler`.

This is largely a distilled re-implementation of parts of
`ConnectionManager` and 'PuppetLanguageClient` and requires the
stdio and tcp implementations to work.