(GH-335) Add option to disable language server
This commit surfaces an option to disable loading the Puppet Language
Server, while still allowing the features that do not require a language
server. The first time this option is set, the user will be presented
with a message explainging what they turned off with an option to not
show this message again. The message won't be shown next time if the
user presses the button.