(GH-439) Remove RestartSession Command
Removes the RestartSession Command from the extension. During the
refactor in https://github.com/lingua-pupuli/puppet-vscode/pull/425 the
ability to restart the session using a command was not retained. After
the refactor it was clear that way it was architected that it would not
allow the restart to work. Considering the 'Reload Window' workflow is
more familiar to users, it was decided to remove it and instead guide
users to reloading the window to restart the extension since state is
kept during a reload.