(GH-459) Remove legacy debug adapter code
Now that the puppetAdapterExecutableCommand has been removed, all of the code
required to implement the command is no longer required.  This commit removes this
code and any upstream dependencies.

This commit also removes the Debug Adapter and dependant code as the new API
makes this redundant.  All of the ruby detection can now use the same file
whereas previously it had to be split up due to different node environments.