(GH-454) Fix PDK Ruby file path
This commit updates the default Ruby version from 2.4.0 to 2.5.1 used
when using PDK as the source for the Puppet Language Server. Ruby 2.5.1
is in PDK 1.7.1 and PDK 1.8.0 so it is safe to use as default until we
can land a feature to allow the user to pick between available PDK ruby
versions.