(GH-469) Restore puppet palatte commands
When adding support for `Puppetfile` as a language it changed the
behavior of `resourceLangId` with contributed commands. This hid any
commands with `when` clauses that use `resourceLangId` to determine
their availability. Changing the `when` clauses to use `editorLangId`
has restored these commands in both the command palatte and in the
editor menus.