(GH-480) Automaticaly find ruby version from PDK
This commit changes the hardcoded paths for Ruby and Gem versions
provided by the PDK to methods that match based on `latest` versions.

This is rather naive sort approach, and it is kinda what got us to
hardcode things in 7b9083e86253d5d5cef540fe4a6f9e39d903f75e
However, the logic works in that the versions are sorted in descending
order this time around instead of assuming the one we want is the last
in the list. The sorted 'first' is our latest, and the one we should use
both for ruby version as well as for the cached version.

This sort can fail if there are weird version numbers that don't match
semver, but ruby generally follows these and Puppet doesn't vendor
interim versions so PDK doesn't either.