(GH-489) Add PDK installation resolver
Previously all of the logic used to detect PDK information was contatined within
the configuration class.  This commit adds a new PDK Resolver file which
contains classes to detect all of the available ruby instances within a PDK
installation, and detect the available Puppet gem versions for each ruby
instance.