(maint) Fix regex for detecting puppet versions
Previously the regular expression for detecting puppet gem versions would only
work on Windows operating systems as the puppet gems have an architecture
suffix e.g. x64.  However on Linux/Mac, the gems do not have this. This commit
updates the filtering regex to look for a suffix or an end of string.