(GH-530) Enable extension on workspace load
This commit updates the `activationEvents` array to activate the
extension if there are any Puppet manifests or module metadata.json
files present. This results in the extensions starting when a folder
containing Puppet code is opened, not only after Puppet code is opened.
This is a better user experience, as it enables the features a user
would want to use without having to have them know to open a Puppet file
first.