(maint) Use file picker for PDK New Module
This commit uses the builtin file picker to choose the folder to create
a new module with PDK. The previous UI required the user to type out the
fully qualified path with no autocomplete, resulting in an error prone
user experience. This presents the user with a platform native file
picker, which then passes a fully qualified path to our methods.