(GH-541) Check for latest PDK Version on Start
This commit compares the installed PDK version with the latest version
published by Puppet in their S3 file using a http request. A setting was
added to allow the user to choose whether or not to check for an update.

This does not attempt to update the PDK, only checks to see if it is
current.