(maint) Convert gulpscript to psakefile
This commit converts the build script from gulp to a psake. The
`psakefile.ps1` now contains all build logic, and comes with a wrapper
script called `build.ps1`. The `build.ps1` script calls `psakefile.ps1`,
ensures psake is installed, and then runs psake with the correct task.

The build process has remained the same, with the following changes:

- Updated npm commands to point to the build.ps1 script
- Removed gulp and all gulp dependencies from devDependencies

A breaking change was added to move the editor-services release
information from editorComponents.json to package.json. This removes the
need for a seperate file to maintain, and keeps all version information
for components in one place. Changes to the key names were done to make
it easier to parse.