(GH-515) Integrate PDK Ruby API
This commit uses the PDK Ruby API instead of the pdk shell commands to
invoke the following PDK commands:

- [x] PDK New Class
- [x] PDK New Defined Type
- [x] PDK New Task

It streamlines the invocation of the following PDK commands, but still
uses the intergrated shell to invoke them:

- [x] PDK New Module
- [x] PDK New Test Unit
- [x] PDK New Validate

It changes the `Pdk New Module` command to use the current workspace
directory if no directory is specified in the quickpick.

It also change the command id of all PDK commands to start with `puppet`
instead of `extension` to standardize on a unique command id namespace.

It fixes the command registration for the PDK commands in the command
palette and editor context menus.