(GH-579) Fix build script extract path
This commit makes the editor-services extract process three steps
instead of one. The first is extracting to a `tmp` folder, then moving
to the correct final folder, then deleting the `tmp` folder.

Previously `Extact-Archive` used the file name of the zip file to
determine the target folder name. This folder name is incorrect.