(613) Puppet Node Graph Feature Rewrite
This commit replaces the viz.js Node Graph Vizualizer with a
cytoscape.js based version. All existing command ids and workflow were
kept, only the presenation logic was changed.

The cytoscape.js module is a full featured data vizualizer that allows
us to display the puppet node graph in a multitude of formats and
presentations. While this commit only replaces existing functionality,
cytoscape has alot of potential for displaying the node graph in
interesting formats since it was designed to handle very large
scientific data sets.

The latest version of the viz.js module changed it's api to a pattern
we could not use reliably in the async language server call, and the
proejct itself has been archived in github and there hasn't been a new
version in several months.

w