(GH-583) Change build script to github format
This commit changes the part of the buildscript that vendors the
specified release of the puppet-editor-services project when a custom
git ref is used, which is a branch on a github remote to use.

Previously the config was:

```
"editorServices": {
  "githubuser": "glennsarti",
  "githubrepo": "puppet-editor-services",
  "githubref": "spike-rearch-langserver"
}
```

This has been changed to:

```
"editorServices": {
  "githubrepo": "puppet-editor-services",
  "githubref": "glennsarti:spike-rearch-langserver"
},
```

This allows someone to copy paste the githubref from the Github page or
the terminal into the config to user without extra changes.

This also streamlines the code that parses the config into a less
complicated statement block.