(GH-644) Consolidate README
This commit consolidates the README and README_BUILD documents into one
README that is suitable for packaging into the extensions as well as
being a landing page for the extension in github.

It also refreshes the README content by removing most of the feature
list and table of contents sections, reodering the sections to surface
troubleshooting and installation needs first and then lists how to
develop the extension.