(GH-649) Reduce Activation Events
This commit removes unneeded activation points from the activationEvents
field in the package.json.

It was reported that the extension is activating at times when the user
does not want it to, removing uncessary ones is the solution.

The Bolt commands would trigger activation of the extension when the
command palate is used, even when there isn't any content in the
workspace to warrant it.

The PDK Commands are only able to be used if inside a puppet module. The
core activation events will cover making these commands available, so
they can be removed.