(maint) Adds tool to update puppet-vscode projects
This commit adds a helper script to synchronise (one-way) release projects in
the puppet-vscode repo.  Editor Services and Editor Syntax projects can also
be used if added to the Project description e.g.

editor-services: <a version>
editor-syntax: <a version>

Any notes etc. that are not in the correct milestones are automatically removed.