(GH-666) Puppetfile view
Adds a new view to the Puppet toolbar view container that shows the puppet modules listed in the Puppetfile in the current workspace. This also adds a new command and context menu that will open the Puppetfile at the line of the selected module in the view.

This uses the `openTextDocument` VS Code API to open the Puppetfile without showing the file to the user. This initiates a `didOpen` notification which sends the content to the Language Server. If we opened the document and sent the content ourselves, a `didOpen` notitication is still sent, so it would result in sending the content twice.