(GH-670) Guard installType and installDirectory
Add more explanation to the settings for installDirectory and installType using new VS Code Settings API that explains when to use each setting. Add a guard clause early on in the activation process that warns when settings are used that are incompatible.

The extension works when there is both pdk and agent present, and when only one is present. The problem here is that the settings for the extension are not helping the user choose successful configurations and allowing mixing of settings that then produce incorrect paths to the parts of puppet we need.

The intention for auto was to not use installDirectory or installType, so that the extension would choose whether to use pdk or agent and which paths to use itself. This was intended to make the default 'happy path' the situation where a user has pdk installed to the default path and not require the user to set any settings.