(maint) Use local tsc instead of global
Previously the build script used a global install of typescript instead of the version installed using npm. This version mismatch means new typescript rules or features may break the build. This changes the build process to use the local version with `npm run compile` so that the build script operates the same as running the project inside VS Code.