(GH-724) PDK New Module with Native Window Picker
This commit updates the PDK New Module command to use the native window picker instead of asking the user to enter in the fully qualified path. It also opens the newly created module inside the window the command was issued in, instead of opening a new VS Code window.

By using the native window picker, we get the correct experience on WIndows, Mac and Linxu. Even inside WSL or even Remote Containers.