(GH-759) Harden PCT content journey
This commit hardens the PCT Create journey adding handling for a
few scenarios where the user might receive un expected exceptions.

The pctExecHelper will now log an error and re-throw any exceptions
that occur when using child_process.exec leaving it to the consumer to
handle as they wish.

Template retrieval will now return an empty array if there has been an
expected exception or if there are no templates to retrieve. In both
cases the user will be appropriately informed via a toast message. The
consumer can then handle this and gracefully terminate the journey.

Finally, extra handling has been added to createContent to appropriately
notify the user when a template name has not been specified or when
the execution of pct new fails.