(bug) - Fix module hover in metadata.json
This commit fixes the module hover in metadata.json. Before it was
returning just `puppetlabs` i.e. the module namespace, as oppose
to the entire module name for example `puppetlabs/stdlib`.

Now, this calculates the position and returns the entire string, which
can they be used to query the forge api for module data.