(CAT-1669) - Fix hashrocket allignment on completion item select
Prior to this commit, when selecting a param/prop from the dropdown
completion list in puppet-vscode, the hash rocket would not be formatted
inline with the rest of the document.

This commit implements a middleware to the language client, and adds a
vscode API item command which allows us to execute a vscode command
after the selection of the param/prop from the dropdown list. So when an
param/prop is selected, we send a callback event to the language server
to format the document.

This is only possible on the vscode side thanks to the vscode API,
unfortunately there is no suitable events to implement this into the
langauge server.